backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
python-dotenv==1.0.0
Pillow==10.0.0
# face-recognition==1.3.0
# dlib==19.24.2
numpy==1.24.3
opencv-python==4.8.0.76
Werkzeug==2.3.7